adds a blog context which contains the functionality of loading posts and uses the context api to pass down this functionality
