+++ src/App.jsx
@@ -1,15 +1,18 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./pages/Home";
 import BlogDetail from "./pages/BlogDetail";
+import { BlogProvider } from "./components/BlogContext";
 
 function App() {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/posts/:postId" element={<BlogDetail />} />
-      </Routes>
-    </BrowserRouter>
+    <BlogProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/posts/:postId" element={<BlogDetail />} />
+        </Routes>
+      </BrowserRouter>
+    </BlogProvider>
   );
 }
